+++ compiler-tests/src/test/data/diagnostic/dependencygraph/DoNotSuggestSimilarBindingSubtypesForAny.kt
@@ -1,0 +1,9 @@
+// RUN_PIPELINE_TILL: FIR2IR
+// RENDER_IR_DIAGNOSTICS_FULL_TEXT
+
+@DependencyGraph
+interface AppGraph {
+  @Provides fun provideAny(): Any = Any()
+  val any: Any
+  val <!METRO_ERROR!>int<!>: Int
+}